backend/face_recognition_service/requirements.txt
flask==2.0.1
flask-cors==3.0.10
face_recognition==1.3.0
numpy==1.21.0
opencv-python==4.5.3.56
pillow==8.3.1